backend/requirements.txt
fastapi==0.115.12
openai==1.79.0
python-dotenv==1.1.0
sqlalchemy==2.0.41
uvicorn==0.34.2